requirements.txt
Flask
flask-jwt-extended
pandas
requests
gunicorn